tools/c7n_salactus/requirements.txt
argparse>=1.2.1
setuptools-scm==1.15.0
arrow>=0.8.0
boto3>=1.4.1
botocore>=1.4.65
click==6.7
docutils>=0.12
Flask>=0.11.1
functools32==3.2.3-2
futures==3.1.1
ipaddress>=1.0.17
itsdangerous==0.24
Jinja2>=2.8
jmespath>=0.9.0
jsonschema>=2.5.1
MarkupSafe==1.0
python-dateutil>=2.5.3
PyYAML>=3.11
redis>=2.10.5
rq-dashboard>=0.3.7
rq>=0.6.0
s3transfer>=0.1.9
six==1.10.0
skew==0.16.1
supervisor>=3.3.2
Werkzeug>=0.11.11
wsgiref==0.1.2
procname==0.3
msgpack-python>=0.4.8
lz4==0.9.0
